requirements.txt
asgiref==3.2.7
Django==3.0.5
django-elasticsearch-dsl==7.1.1
django-elasticsearch-dsl-drf==0.20.8
django-filter==2.2.0
django-haystack==2.8.1
django-nine==0.2.3
djangorestframework==3.11.0
elasticsearch==7.6.0
elasticsearch-dsl==7.1.0
Pillow==7.1.1
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
